+++ module/LearnZF2FormUsage/src/LearnZF2FormUsage/Controller/FormController.php
@@ -15,6 +15,7 @@
  * This software consists of voluntary contributions made by many individuals
  * and is licensed under the MIT license.
  */
+
 namespace LearnZF2FormUsage\Controller;
 
 use Zend\Form\FormInterface;
@@ -29,7 +30,7 @@ class FormController extends AbstractActionController
     protected $form;
 
     /**
-     * Construct form property
+     * Construct form property.
      *
      * @param FormInterface $form
      */